Refactor for loop in main function
Remove text variable and put command line argument directly in for statement as args.text.

Refactor conditional if-else statement, using the dict.get method to print the key-value. Set the default for jumper.get() to print char if value not in dictionary.

Add final print() function outside for loop so that the entire text output of the for loop is printed on new line in terminal.