Add text input and file output arguments.
Add code to get_args() that opens and reads a file if included in arguments. Strip off trailing space and change args.text variable. This handles the input from the user if it is a file and writes it to args.text variable. Then in main() args.text is all ready handled wherever the argument text comes from.

When run, it has all ready been handled somewhere else.