Refactor to simplify and remove error chances.
Since both conditional statements were printing the same output, it is best to move the print line out of the conditional branch, and create the uppercase output there.

The variable out_fh is assigned None so that if it is not empty, and outfile is True, use stdout from sys module instead, assigning it to out_fh and printing that in uppercase.