Refactor as low memory version
Use IO module to change args.text from a string. If that string names an actual file on disk, then open that file and args.text will be some file.

Else if it's from the command line wrap it in StringIO and pretend its a file handle.

In main(), iterate over lines in file. This makes it read one line at a time, so we are not allocating all the memory to read it at one time.

Remove '\n' in for loop in main(), since now we're iterating over lines of text, the new line is still attached. Adding a new line would add an extra line that does not belong.