Add optional nargs and make default to stdin
In args.file argument, nargs='*' allows for 0 or more arguments (not required, optional).

If an argument is provided, it needs to be a readable file. If no argument is provided, the default will be stdin. Reason for making it a list is that nargs will make it a list, so the sys.stdin needs to match.