Format output with f string
Created a loop to iterate over each fine in the fh input. Counted and added for each variable. Printed the result using f string, outside the loop, with :8 at the end of the variable to put in spaces for formatting.

Created tuple with totals of each line, word, and byte, and added to that from the loop variables. If the len of the FILE input is more than one, print the totals along with the word total.