Create flag options to count elements
Use single letter flags in arguments to count lines, words, or characters, displaying each specified item. If any of the flags are present, in main(), print the count of the flagged variable, adding an if statement on the end to only print it if the flag is specified. Else print an empty string, consistent with the default/main parts of the program.