Initial commit
Add choices to vowel argument to automatically print a message with vowels if anything else is entered into the input.

Create regular expression with re module. Use match object and sub() method to replace matching instances of reg_ex with the vowel input from the command line. Defaults to 'a'.