Refactor file handler
Add file handler argument to argparse to handle 'wt' file, and sys.stdout if no --outfile flag. Makes the code much easier to read, understand, and work with. Faced a difficulty with printing when handled this way, but in the print() function, argument needs to be file=args.outfile to make it work.

Used for loop inside a list variable verses to call the verse() function for each day.