Create function to break word
Stemmer() takes word input from args.word and returns the word in two parts. If the word begins with a consonant, the function returns the consonant(s) and the rest of the word.

If the word begins with a vowel, the function returns the word in the form of a tuple with an empty string as the first element. The second element is the word. If the word has no vowels, the function returns the word in the form of a tuple with the word as the first element and an empty string as the second element.