Write main() function
Use list comprehension to create a list of the second return value of stemmer() function. Since the return value was a tuple, we need to access the second element of the tuple. The second element is the list of stemmed words with each of the letters (consonants + clusters) prepended on the word, unless the letter is the same as the first return value(the first letter of the word).

Print the list of words sorted with a new line between each word and sorted in alphabetical order. If the return value from stemmer does not have any vowels in either element, print "Cannot rhyme "{args.word}"' instead.