Refactor main function
Combined the prefixes and letters into a single variable and split the string of consononant clusters at the space.

If rest, the value of the second tuple, is present, then use a list comprehension to create a list of the consonant clusters joined with the rest of the word, except if the start letter is the same.

Else print error message.