Use regular expression to capture and replace text in a file or input
Add for loop to go through each line of args.text, splitlines() is used to split the text into lines and store them in a list. FOr each item in the list, use re.split to split the lines into words and nonword characters using (\W+). Strip each line and use the fry function on each word, append them to the list, and print the joined words.