Initial commit
This is the initial commit for the mad libs project. This commit
adds the mad.py file which contains the code for the mad libs
game. The game is also described in the README.md file.

The mad libs game is a simple game in which the user is prompted
to enter a series of words. The words are then used to fill in
placeholders in a story. The story is then printed out to the
console.

Changes to be committed:
	new file:   17_mad_libs/mad.py