Replace regex matches with user input
Loop through the matches and unpack the tuples (placeholder, name) for each match. Then ask the user for a replacement for the placeholder, if the inputs are not given with args.input.

If inputs are entered as argument, pop the first value from the list of inputs and use it as the replacement.

Text will be reassigned to the result of the re.sub() call, which replaces the placeholder with the replacement.