Use the csv module to read the csv file and return a list of tuples.
The csv module is part of the standard library and is used to read and write csv files. The csv module has a reader function that takes a file object and returns an iterable object that can be used to iterate over the rows of the csv file. Each row is returned as a list of strings. The csv module also has a DictReader class that returns an iterable object that can be used to iterate over the rows of the csv file. Each row is returned as a dictionary where the keys are the column names and the values are the values in the row.

I changed the return value of the read_csv function to a list of tuples. The reason for this is that the csv module does not support reading a subset of the columns in the csv file. The csv module will read the entire csv file and return all of the data. The list of tuples allows us to return only the data we need.

I needed to change the reps range from a string to an integer. The csv module returns all of the data as strings. I used a regular expression to match each group of 1 or more digits and capture it as a group, and I used the int function on each captured group to convert the string to an integer.

Finally I appended the name of each exercise, the integer value of the lowest rep, and the integer value of the highest rep to the list of tuples.

I forgot to put () around the appended items at first, which gave me the error of: 'cannot append more than 1 item to a list'.