Refactor source code
Fixes #1
Use map function when assigning low, high in read_csv to split the string at the '-' character, and convert the strings to integers using int() function, the first map argument.

Fixes #2
Initiate empty list as wod. Use a foor loop of random.sample(exercises, ...)  which creates a list of tuples. Assign 3 variables for each iteration of the loop: exercise, low, and high.

Pick a random number between low and high, and append the tuple to wod. If the easy flag is given as argument, divide reps by two. Use the int() function to convert the result to an integer.